Add new 01-Login project
Deprecating the 00-Seed-Project tutorial in favor of a more up-to-date
installation and documentation. This addition will be referenced by an
updated Laravel QS document. This commit adds a deprecation notice on
the 00-Seed-Project readme and clarfies install steps for Docker and
Heroku by walking through both processes start to finish.